refactor(search): split sidebar search support
